requirements.txt
azlyrics==1.3.2
beautifulsoup4==4.9.0
billboard.py==6.1.0
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
Flask==1.1.2
Flask-SocketIO==4.3.1
gevent==20.5.0
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
lyricsmaster==2.8.1
lyricwikia==0.1.11
MarkupSafe==1.1.1
matplotlib==3.2.1
murmurhash==1.0.2
nltk==3.3
numpy==1.18.4
oauthlib==3.1.0
pendulum==2.1.0
plac==1.1.3
preshed==3.0.2
pymusixmatch==0.3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-engineio==3.13.1
python-socketio==4.6.0
pytzdata==2019.3
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
spacy==2.2.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
spotipy==2.12.0
srsly==1.0.2
stem==1.8.0
thinc==7.4.0
tqdm==4.46.0
twython==3.8.2
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
zipp==3.1.0
